Refactor a lot of build things for icons
