+++ .github/workflows/release.yml
@@ -28,7 +28,3 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
         run: npx semantic-release
-      - name: Eik login and publish
-        run: pnpm eik login -k $EIK_TOKEN && pnpm eik publish || true
-        env:
-          EIK_TOKEN: ${{ secrets.EIK_TOKEN }}
