+++ scripts/output/react.js
@@ -27,7 +27,8 @@ getSVGs().forEach(({ svg, filename, exportName, name }) => {
     `import { activateI18n } from '../src/utils/i18n';`,
     `activateI18n(enMessages, nbMessages, fiMessages);`,
     `const title = i18n.t({ message: \`${message}\`, id: '${id}', comment: '${comment}' });`,
-    `export const ${exportName} = (attrs) => React.createElement('svg', { ${attrs.join(", ")}, dangerouslySetInnerHTML: { __html: ${'`'}${titleHtml}${svg.html}${'`'} }, ...attrs, });`,
+    `const ${exportName} = (attrs) => React.createElement('svg', { ${attrs.join(", ")}, dangerouslySetInnerHTML: { __html: ${'`'}${titleHtml}${svg.html}${'`'} }, ...attrs, });`,
+    `export default ${exportName};`
   ].join("\n");
 
   const path = joinPath(basepath, filename)
